requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==23.2.0
boto3==1.36.10
botocore==1.36.10
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
dateparser==1.2.1
dill==0.3.9
docker==7.1.0
exceptiongroup==1.2.2
Farama-Notifications==0.0.4
fastapi==0.115.8
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
gym==0.26.2
gym-anytrading==2.0.0
gym-notices==0.0.8
gymnasium==1.0.0
h11==0.14.0
h5py==3.13.0
idna==3.10
importlib-metadata==6.11.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
mock==4.0.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pandas_ta==0.3.14b0
pathos==0.3.3
pillow==11.1.0
platformdirs==4.3.6
pox==0.3.5
ppft==1.7.6.9
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
pyarrow==19.0.0
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.0
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.11.2
sagemaker==2.239.0
sagemaker-core==1.0.21
schema==0.7.7
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
stable_baselines3==2.5.0
starlette==0.45.3
sympy==1.13.1
tblib==3.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
torch==2.6.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==2.1.3
yarl==1.18.3
zipp==3.21.0
